Changed ShopManager.java to: class responsible for taking user's input and validating if it's correct UserInputProvider class that takes int input and basing on it makes a decision what to do ShopService class that wraps it all together in a simple while loop executing one then another thing in it, meaning that it will use 1st class to take input, then it will pass it to second to execute what needs to be done ShoppingJourney
