+++ src/main/java/Main.java
@@ -1,10 +1,10 @@
-import basket.factory.ShopManager;
+import basket.factory.ShoppingJourney;
 
 
 public class Main {
     public static void main(String[] args) {
         System.out.println("Welcome");
-        new ShopManager().shopManagerMenu();
+        new ShoppingJourney().shopManagerMenu();
         System.out.println("End of shopping. See you again");
     }
 }
